preview: pass the path to the file
Passing file contents to the renderer instead of a path
is extremely unwise. Encoding the entire content each time,
even though you already have a file output from qemu, is at
the very least very unreasonable. The author (me) was probably
not in his right mind when writing this code.